put apiMiddleware.ts in APIRequestHandling folder
