+++ src/client/Components/LanguagePage/LanguagePage.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react'
 import { connect } from 'react-redux'
 import { ReduxState } from '../../store'
 import { RouteComponentProps } from 'react-router-dom'
-import { languagePageDataRetrivalThunk } from '../../store/APIRequestsHandling/thunks'
+import { languagePageDataRetrivalThunk } from '../../store/APIRequestsHandling/multiplereducerthunks'
 import { getUsersOfLanguageInformation } from './helperfunctions'
 import { IUsersofLanguageInformation } from './shared-types'
 import { IUserReducerDataSlice } from '../../store/user/reducer'
