+++ src/client/Components/LoggedInUserController.tsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux'
 import { Route, Switch, Redirect } from 'react-router'
 import Home from './Home/'
 import LanguagePage from './LanguagePage'
-import { userLoggedInThunk } from '../store/APIRequestsHandling/thunks'
+import { userLoggedInThunk } from '../store/APIRequestsHandling/multiplereducerthunks'
 import { User } from '../../entities'
 import { ReduxState } from '../store'
 import { IUserReducerDataSlice } from '../store/user/reducer'
